game switcher test
Use START button long press for activate game switcher. In switcher, use LEFT/RIGHT to select game, A to reload game, B to return to menu.

+++ .tmp_update/scripts/principal.sh
@@ -9,15 +9,22 @@ runifnecessary(){
 
 rm /mnt/SDCARD/.tmp_update/flags/.save_active
 while [ 1 ]; do
-    # create in menu flag
-    touch /mnt/SDCARD/.tmp_update/flags/in_menu.lock
 
-    runifnecessary "keymon" ${SYSTEM_PATH}/app/keymon
-    cd ${SYSTEM_PATH}/app/
-    ./MainUI  &> /dev/null
+    if [ -f /mnt/SDCARD/.tmp_update/flags/gs.lock ] ; then
+        /mnt/SDCARD/.tmp_update/scripts/gameswitcher.sh
+    fi
+    
+    if [ ! -f /tmp/cmd_to_run.sh ] ; then
+        # create in menu flag
+        touch /mnt/SDCARD/.tmp_update/flags/in_menu.lock
 
-    # remove in menu flag
-    rm /mnt/SDCARD/.tmp_update/flags/in_menu.lock
+        runifnecessary "keymon" ${SYSTEM_PATH}/app/keymon
+        cd ${SYSTEM_PATH}/app/
+        ./MainUI  &> /dev/null
+
+        # remove in menu flag
+        rm /mnt/SDCARD/.tmp_update/flags/in_menu.lock
+    fi
 
     if [ -f /tmp/.cmdenc ] ; then
         /root/gameloader
